Add new modules sed_kernels.py
