+++ diffsky/experimental/kernels/constants.py
@@ -1,0 +1,3 @@
+""""""
+
+LGMET_SCATTER = 0.2
