+++ diffsky/experimental/kernels/dbk_kernels.py
@@ -174,6 +174,29 @@ def _get_dbk_phot_from_dbk_weights(
     return obs_mags_bulge, obs_mags_disk, obs_mags_knots
 
 
+@jjit
+def _get_dbk_linelum_decomposition(dbk_weights, spec_kern_results, ssp_data):
+    linelums_bulge = sspwk._compute_linelum_from_weights(
+        jnp.log10(dbk_weights.mstar_bulge),
+        spec_kern_results.dust_ftrans_lines,
+        ssp_data,
+        dbk_weights.ssp_weights_bulge,
+    )
+    linelums_disk = sspwk._compute_linelum_from_weights(
+        jnp.log10(dbk_weights.mstar_disk),
+        spec_kern_results.dust_ftrans_lines,
+        ssp_data,
+        dbk_weights.ssp_weights_disk,
+    )
+    linelums_knots = sspwk._compute_linelum_from_weights(
+        jnp.log10(dbk_weights.mstar_knots),
+        spec_kern_results.dust_ftrans_lines,
+        ssp_data,
+        dbk_weights.ssp_weights_knots,
+    )
+    return linelums_bulge, linelums_disk, linelums_knots
+
+
 DBKWeights = namedtuple(
     "DBKWeights",
     (
