+++ diffsky/experimental/kernels/tests/test_constants.py
@@ -1,0 +1,7 @@
+""""""
+
+from .. import constants
+
+
+def test_constants_has_expected_globals():
+    assert constants.LGMET_SCATTER == 0.2
