+++ diffsky/experimental/kernels/tests/test_sed_kernels.py
@@ -1,0 +1,7 @@
+""""""
+
+from .. import sed_kernels
+
+
+def test_sed_kern():
+    pass
